refactor(confirm-modal): rename
