+++ packages/bezier-react/src/components/Modals/ConfirmModal/ConfirmModal.stories.tsx
