+++ packages/bezier-react/src/components/Modals/ConfirmModal/ConfirmModal.types.ts
