+++ packages/bezier-react/src/components/Modals/ConfirmModal/ConfirmModalHelpers.tsx
